Extract BeaconBlockBodySchema interface
