+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/schemas/SchemaDefinitionsPhase0.java
@@ -15,6 +15,7 @@ package tech.pegasys.teku.spec.schemas;
 
 import tech.pegasys.teku.spec.constants.SpecConstants;
 import tech.pegasys.teku.spec.datastructures.blocks.blockbody.BeaconBlockBodySchema;
+import tech.pegasys.teku.spec.datastructures.blocks.blockbody.BeaconBlockBodySchemaPhase0;
 import tech.pegasys.teku.spec.datastructures.state.beaconstate.BeaconStateSchema;
 import tech.pegasys.teku.spec.datastructures.state.beaconstate.versions.phase0.BeaconStateSchemaPhase0;
 
@@ -31,7 +32,7 @@ public class SchemaDefinitionsPhase0 implements SchemaDefinitions {
   }
 
   @Override
-  public BeaconBlockBodySchema getBeaconBlockBodySchema() {
-    return BeaconBlockBodySchema.create(specConstants);
+  public BeaconBlockBodySchema<?> getBeaconBlockBodySchema() {
+    return BeaconBlockBodySchemaPhase0.create(specConstants);
   }
 }
